[Tests] Abstract tier2 init in PivxTier2TestFramework
